Add retain_mut (#198)
* Add retain_mut

This adds a `retain_mut` method to `ArrayVec` and
`TinyVec` that's just `retain` but with mutable
references. This method is present on
`std::vec::Vec` as of a while ago so it's nice to
have in `tinyvec` too.

Fixes #195

* `TinyVec::retain_mut` requires rustc_1_61 feature

It calls `std::vec::Vec::retain_mut` which is only
available since Rust 1.61. People that still want
to use a very old compiler (CI seems to use 1.47)
still get to compile the crate, they just don't
get `TinyVec::retain_mut`.

Note that `ArrayVec::retain_mut` is implemented
inside the crate so it's still available, just the
`TinyVec` format is not.

* Add some tests for ArrayVec::retain_mut