Add ArrayVec::as_inner() (#197)
Originally I was going to call this `as_array()`
but `as_inner()` is more consistent with the
existing `into_inner()` method. I don't
_ultimately_ care what it ends up being called.

I wonder if we should have a `impl
From<ArrayVec<T>> for A` implementation or is that
a bit too foot-gun-y?

This lives in its own `impl<A> ArrayVec<A>` block
because rust 1.47.0 is unhappy about having a
const function inside `impl<A: Array>
ArrayVec<A>`. We don't need the trait anyway.

Fixes #193.