Fix rustc features not depending on each other
This crate has several features related to the rustc version. It is currently a bug that not all off these features depend on the lower rustc version features. This commit fixes this.

This bug leads to bad user experience. For example, you notice that by default tinyvec doesn't work with arrays of all sizes. The docs tell you to enable the rustc_1_55 feature to fix this. You see that there are features for even newer rustc versions so you pick the most recent one because you use the current stable compiler. To your surprise, this does not enable tinyvec to work with arrays of all sizes.

This change is correct because features are additive and rustc versions are backward compatible.