Make Debug instances behave like those for slices (#211)
The indentantion in case of structs inside the
arrays was wrong with pretty debug instance.
Instead of rolling one by hand, just use the
underlying slice implementation. It appears that
this was even the intention orginally, based on
the `TinyVec_pretty_debug` test. The main issue is
that the test used a simple value instead of a
struct so it missed the problem. FWIW I have kept
the existing test but I think it could just be
deleted now.

I haven't touched the `Display` version because
while slices have a `Debug` impl, there isn't one
for `Display`. So I think probably `Display` has
the same problem.

These tests can technically be written in a
non-alloc/non-std way by defining an array-backed
buffer and giving it `core::fmt::Write` but I was
too lazy to do so: it doesn't seem useful.

Fixes #192.